refactor: 생성/수정 날짜를 추상클래스로 분리
