+++ src/main/java/cholog/wiseshop/api/order/dto/CreateOrderRequest.java
@@ -1,0 +1,5 @@
+package cholog.wiseshop.api.order.dto;
+
+public record CreateOrderRequest(Long productId,
+                                 int count) {
+}
