+++ src/main/java/cholog/wiseshop/api/order/service/OrderService.java
@@ -1,0 +1,20 @@
+package cholog.wiseshop.api.order.service;
+
+import cholog.wiseshop.db.order.OrderRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class OrderService {
+
+    private final OrderRepository orderRepository;
+
+    public OrderService(OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
+
+    public Long createOrder() {
+
+    }
+}
